GDMS Schema v2.1 (GCP-8447, GCP-8448, GCP-8449, GCP-8450, GCP-8451, GCP-8452, GCP-8453, GCP-8454, GCP-8455, GCP-8456, GCP-8459, GCP-8469, GCP-8506)

+++ src/main/java/org/generationcp/middleware/pojos/gdms/QtlDetailElement.java
@@ -128,11 +128,11 @@ public class QtlDetailElement implements Serializable{
         qtlDetails.setRightFlankingMarker(rightFlankingMarker);
     }
     
-    public Integer getEffect() {
+    public Float getEffect() {
         return qtlDetails.getEffect();
     }
     
-    public void setEffect(Integer effect) {
+    public void setEffect(Float effect) {
         qtlDetails.setEffect(effect);
     }
